refactor: move soy.js into utils/
As in the previous commit, the motivation here is to keep only scripts
corresponding to subcommands that you can run with "liferay-npm-scripts
subcommand" in the "src/scripts/" directory, to avoid confusion.
