+++ packages/liferay-npm-scripts/src/scripts/build.js
@@ -10,7 +10,7 @@ const spawnSync = require('../utils/spawnSync');
 const fs = require('fs');
 const path = require('path');
 
-const {buildSoy, cleanSoy, soyExists} = require('./soy');
+const {buildSoy, cleanSoy, soyExists} = require('../utils/soy');
 const {removeBabelConfig, setBabelConfig} = require('../utils/babel');
 const getMergedConfig = require('../utils/get-merged-config');
 const {moveToTemp, removeFromTemp} = require('../utils/move-to-temp');
